requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
oauthlib==3.1.0
pyenv-win==1.2.4
PySocks==1.7.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.8.0
urllib3==1.25.9